requirements.txt
amqp==5.1.1
asgiref==3.5.2
autopep8==2.0.0
billiard==3.6.4.0
boto3==1.26.27
botocore==1.29.27
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.4
django-amazon-ses==4.0.1
django-braces==1.15.0
django-celery-beat==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
drf-tracking-logger==1.9.3
firebase-admin==6.0.1
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
jmespath==1.0.1
jwcrypto==1.4.2
kombu==5.2.4
more-itertools==9.0.0
msgpack==1.0.4
oauthlib==3.2.2
phonenumbers==8.13.2
Pillow==9.3.0
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==4.21.11
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycurl==7.45.1
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.11.0